Move file based deployment lock to file state implementation.
